% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basal_df.R
\docType{data}
\name{basal_df}
\alias{basal_df}
\title{Basal test dataframe}
\format{
A data frame with 88 rows and 8 variables:
\describe{
\item{Time}{time test, in minutess}
\item{HR}{heart rate, in beats/min}
\item{VO2}{volume of oxygen consumption, in ml/min}
\item{VCO2}{volume of exhaled carbon dioxide, in ml/min}
\item{RER}{respiratory exchange ratio}
\item{BF}{breathe frequency, in breaths/min}
\item{VE}{ventilation, in l/min}
\item{PETCO2}{end-tidal carbondioxide pressure, in mmHg}
}
}
\usage{
data(basal_df)
}
\description{
A dataframe with the results of a test to assess basal metabolism
}
\keyword{datasets}
