% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_steps}
\alias{calculate_steps}
\title{Calculate steps}
\usage{
calculate_steps(step_time, db, db_type)
}
\arguments{
\item{step_time}{how often the data was collected (in seconds).}

\item{db}{a database}

\item{db_type}{either "basal" or "MFO"}
}
\description{
Calculate steps
}
