% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{calculate_vars}
\alias{calculate_vars}
\title{Calculation of CHO, FAT and Kcal}
\usage{
calculate_vars(step_time, db_MFO, VO2max, author)
}
\arguments{
\item{author}{}
}
\value{

}
\description{
Calculation of CHO, FAT and Kcal
}
