% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_5min}
\alias{get_5min}
\title{Get a 5 minutes database}
\usage{
get_5min(db, cv_var, n_row)
}
\arguments{
\item{db}{a database}

\item{cv_var}{variable to calculate coefficient of variation}

\item{n_row}{number of rows}
}
\description{
Get a 5 minutes database
}
