% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{met_basal}
\alias{met_basal}
\title{Basal metabolic rate}
\usage{
met_basal(step_time, db, cv_var)
}
\arguments{
\item{step_time}{how often the data was collected (in seconds).}

\item{db}{a database}

\item{cv_var}{variable to calculate coefficient of variation}
}
\description{
Basal metabolic rate
}
