% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFO.R
\name{MFO}
\alias{MFO}
\title{Maximal Fat Oxidation Function}
\usage{
MFO(step_time, db_MFO, db_basal, db_graded, cv_var, author, VO2max = NULL)
}
\arguments{
\item{step_time}{every few seconds the data was collected.}

\item{db_MFO}{database containing MFO test.}

\item{db_basal}{database containing basal test.}

\item{db_graded}{database containing incremental exercise test.}

\item{cv_var}{variable to estimate coefficient of variation. Can be: VO2, VCO2 or RER.}

\item{author}{author to estimate MFO. Can be: Frayn or Jeukendrup.}

\item{VO2max}{VO2max can be passed directly using this argument instead of use db_graded argument.}
}
\description{
Maximal Fat Oxidation Function
}
\examples{
\dontrun{
MFO(step_time = 20,
    db_MFO = db_MFO,
    db_basal = db_basal,
    db_graded = db_graded,
    cv_var = "RER",
    author = "Frayn",
    VO2max = NULL)
}
}
