% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFO_kinetics.R
\name{MFO_kinetics}
\alias{MFO_kinetics}
\title{Maximal Fat Oxidation Kinetics}
\usage{
MFO_kinetics(MFO_data)
}
\arguments{
\item{MFO_data}{a data frame obtained from MFO function}
}
\value{

}
\description{
Maximal Fat Oxidation Kinetics
}
