% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFOs.R
\name{MFOs}
\alias{MFOs}
\title{Maximal Fat Oxidation calculation of multiple databases}
\usage{
MFOs(folders_path, step_time, cv_var, author, VO2max = NULL, save_plot = T)
}
\arguments{
\item{folders_path}{path to the folder with the databases}

\item{step_time}{every few seconds the data was collected}

\item{cv_var}{variable to estimate coefficient of variation. Can be: VO2, VCO2 or RER.}

\item{author}{author to estimate MFO. Can be: Frayn or Jeukendrup.}

\item{VO2max}{VO2max can be passed directly using this argument instead of use db_graded argument. Default set to NULL.}

\item{save_plot}{to save the plot or not.Default set to True.}
}
\description{
Maximal Fat Oxidation calculation of multiple databases
}
\examples{
\dontrun{
MFOs(folders_path = /Users/username/Documents/R_packages/MFO
    step_time = 20,
    cv_var = "RER",
    author = "Frayn",
    VO2max = NULL,
    save_plot = T)
}
}
