% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{read_MFO_databases}
\alias{read_MFO_databases}
\title{Title}
\usage{
read_MFO_databases(
  from = c("folder", "files"),
  path,
  db_basal_name,
  db_MFO_name,
  db_graded_name,
  col_name_VO2,
  col_name_VCO2,
  col_name_RER,
  col_name_HR,
  remove_rows = NULL
)
}
\arguments{
\item{db_graded_name}{}
}
\value{

}
\description{
Title
}
