% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{sample_read}
\alias{sample_read}
\title{Read smaple data}
\usage{
sample_read(path)
}
\arguments{
\item{path}{path to sample data}
}
\value{

}
\description{
Read smaple data
}
\examples{
path <- system.file("extdata", "sample_data.xlsx", package = "MFO")
sample_data <- sample_read(path)
}
