% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VO2max_df.R
\docType{data}
\name{VO2max_df}
\alias{VO2max_df}
\title{Graded exercise test dataframe}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 9 columns.
}
\usage{
data(VO2max_df)
}
\description{
A dataframe with the results of a graded exercise test
}
\section{Variables}{

\describe{
\item{Time}{time test, in minutess}
\item{HR}{heart rate, in beats/min}
\item{Load}{Load of the test, in watts}
\item{VO2}{volume of oxygen consumption, in ml/min}
\item{VCO2}{volume of exhaled carbon dioxide, in ml/min}
\item{RER}{respiratory exchange ratio}
\item{VE}{ventilation, in l/min}
\item{PETCO2}{end-tidal carbondioxide pressure, in mmHg}
}
}

\keyword{datasets}
