% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_vars}
\alias{calculate_vars}
\title{Calculation of CHO, FAT and Kcal}
\usage{
calculate_vars(step_time, db_MFO, VO2max, author)
}
\arguments{
\item{step_time}{how often the data was collected (in seconds).}

\item{db_MFO}{dtabase with MFO test}

\item{VO2max}{maximum oxygen uptake}

\item{author}{eithe "Frayn" or "Jeukendrup"}
}
\description{
Calculation of CHO, FAT and Kcal
}
