% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFO_df.R
\docType{data}
\name{MFO_df}
\alias{MFO_df}
\title{MFO test dataframe}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 45 rows and 8 columns.
}
\usage{
data(MFO_df)
}
\description{
A dataframe with the results of a test to assess MFO metabolism
}
\section{Variables}{

\describe{
\item{Time}{time test, in minutess}
\item{HR}{heart rate, in beats/min}
\item{VO2}{volume of oxygen consumption, in ml/min}
\item{VCO2}{volume of exhaled carbon dioxide, in ml/min}
\item{RER}{respiratory exchange ratio}
\item{VE}{ventilation, in l/min}
\item{PETCO2}{end-tidal carbondioxide pressure, in mmHg}
}
}

\keyword{datasets}
