% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFO_kinetics.R
\name{MFO_kinetics}
\alias{MFO_kinetics}
\title{Maximal Fat Oxidation Kinetics}
\usage{
MFO_kinetics(MFO_data)
}
\arguments{
\item{MFO_data}{a data frame obtained from MFO function}
}
\description{
Maximal Fat Oxidation Kinetics
}
\examples{
\dontrun{
#' # Path to the MFO package sample data
path <- system.file("extdata", "sample_data.xlsx", package = "MFO")

# Read databases
sample_data <- read_MFO_databases(from = "files",
                                 path = paste(path),
                                 db_basal_name = "M.BASAL",
                                 db_MFO_name = "MFO",
                                 db_graded_name = "V02máx.",
                                 col_name_VO2 = "V'O2",
                                 col_name_VCO2 = "V'CO2",
                                 col_name_RER = "RER",
                                 col_name_HR = "HR",
                                 remove_rows = NULL)
# Calculate MFO and Fatmax
result_MFO <- MFO(step_time = 20,
                 db_MFO = sample_data$participant_db_MFO,
                 db_basal = sample_data$participant_db_basal,
                 db_graded = sample_data$participant_db_graded,
                 cv_var = "RER",
                 author = "Frayn",
                 VO2max = NULL)
}

}
