% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFOs.R
\name{MFOs}
\alias{MFOs}
\title{Maximal Fat Oxidation calculation of multiple databases}
\usage{
MFOs(
  from = c("folder", "files"),
  path,
  db_basal_name,
  db_MFO_name,
  db_graded_name,
  step_time,
  cv_var,
  author,
  VO2max = NULL,
  remove_rows = NULL,
  col_name_VO2 = "VO2",
  col_name_VCO2 = "VCO2",
  col_name_RER = "RER",
  col_name_HR = "HR",
  save_plot = T
)
}
\arguments{
\item{step_time}{how often the data was collected (in seconds).}

\item{cv_var}{variable to estimate coefficient of variation. Can be: VO2, VCO2 or RER.}

\item{author}{author to estimate MFO. Can be: Frayn or Jeukendrup.}

\item{VO2max}{VO2max can be passed directly using this argument instead of use db_graded argument. Default set to NULL.}

\item{save_plot}{to save the plot or not.Default set to True.}

\item{folders_path}{path to the folder with the databases}
}
\description{
Maximal Fat Oxidation calculation of multiple databases
}
\examples{
\dontrun{

# Path to the MFO package sample data FOLDER
path <- system.file("extdata", package = "MFO")

# Calculate MFO and Fatmax
MFOs(folders_path = path
    step_time = 20,
    cv_var = "RER",
    author = "Frayn",
    VO2max = NULL,
    save_plot = T)
}
}
